pyproject.toml
[project]
name = "madmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "deltachat-rpc-client>=2.37.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "toml>=0.10.2",
]
